SSHOC-34 actor external ids External ids in actors model Add entity for external id Add entity for actor source services Add external id mapping for requests and responses

+++ src/main/java/eu/sshopencloud/marketplace/services/actors/ActorSourceService.java
@@ -64,8 +64,8 @@ public class ActorSourceService {
     }
 
     public void deleteActorSource(String code) {
-//        if (actorSourceRepository.isActorSourceInUse(code))
-//            throw new IllegalArgumentException(String.format("Actor source %s is in use so it cannot be removed anymore", code));
+        if (actorSourceRepository.isActorSourceInUse(code))
+            throw new IllegalArgumentException(String.format("Actor source %s is in use so it cannot be removed anymore", code));
 
         try {
             actorSourceRepository.deleteById(code);
